scala 3 fix part 2
